Title: 사칙연산 함수, Time: 100ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 240801/사칙연산 함수/quadratic-operations-function.py
@@ -8,7 +8,7 @@ def mul(a, c):
     return a * c
 
 def div(a, c):
-    return a / c
+    return int(a / c)
 
 a, o, c = input().split()
